hackerrank v0.1.0 FP.Recursion.Advanced

Elixir solutions for advanced HackerRank functional programming challenges: recursions.

Types

sequences()

sequences() :: [tuple()]

solution()

solution() :: [{char(), tuple()}]

Functions

convex_hull_perimeter(points)

convex_hull_perimeter([tuple()]) :: number()

Convex hull

https://www.hackerrank.com/challenges/convex-hull-fp/problem

Find the perimeter of a convex hull from a series of coordinates, opted to implement Graham's scan algorithm. https://en.wikipedia.org/wiki/Graham_scan

cross_words(grid, words)

cross_words(list(), list()) :: list()

Crosswords-101

https://www.hackerrank.com/challenges/crosswords-101/problem

draw_trees(iterations)

draw_trees(integer()) :: list()

Functions and Fractals - Recursive Trees

https://www.hackerrank.com/challenges/fractal-trees/problem

draw_triangles(iterations)

draw_triangles(integer()) :: list()

Functions and fractals: Sierpinski triangle

https://www.hackerrank.com/challenges/functions-and-fractals-sierpinski-triangles/proble

graham_scan(points)

graham_scan([tuple()]) :: [tuple()]

Graham Scan - find points of a convex hull from a series of coordinates

See https://en.wikipedia.org/wiki/Graham_scan

For https://www.hackerrank.com/challenges/convex-hull-fp/problem

super_digit(n, k)

super_digit(integer(), integer()) :: integer()

Super digit

https://www.hackerrank.com/challenges/super-digit/problem